Revert "Revert "chore: Better organize component folders""
This reverts commit c783701371e673ba533f44e4ba402116489c930d.
